chore: upgrade firebase package and made required changes to getToken and onNewToken
