+++ packages/pushnotification/android/build.gradle
@@ -3,9 +3,10 @@
 buildscript {
     repositories {
         jcenter()
+				google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
+        classpath 'com.android.tools.build:gradle:3.5.1'
         classpath 'com.google.gms:google-services:3.2.0'
     }
 }
@@ -26,11 +27,11 @@ def getExtOrDefault(name, fallback) {
 apply plugin: "com.android.library"
 
 android {
-    compileSdkVersion getExtOrDefault('compileSdkVersion', 28) 
-    buildToolsVersion getExtOrDefault('buildToolsVersion', "28.0.3") 
+    compileSdkVersion getExtOrDefault('compileSdkVersion', 28)
+    buildToolsVersion getExtOrDefault('buildToolsVersion', "28.0.3")
     defaultConfig {
-        minSdkVersion getExtOrDefault('minSdkVersion', 16) 
-        targetSdkVersion getExtOrDefault('targetSdkVersion', 28) 
+        minSdkVersion getExtOrDefault('minSdkVersion', 16)
+        targetSdkVersion getExtOrDefault('targetSdkVersion', 28)
         versionCode 1
         versionName "1.0"
         ndk {
@@ -50,7 +51,7 @@ dependencies {
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
-    compile 'com.google.firebase:firebase-messaging:12.0.1'
-    compile 'com.google.firebase:firebase-core:12.0.1'
-    compile 'com.firebase:firebase-jobdispatcher:0.6.0'
+    compile 'com.google.firebase:firebase-messaging:23.0.0'
+	compile 'com.google.firebase:firebase-core:20.0.0'
+	compile 'com.firebase:firebase-jobdispatcher:0.6.0'
 }
